Add the ability to have the measure display units be different than the units used in the point cloud
The setLengthUnit() method that existed before only set the labels that display when measurements are being made.  This is great when the data happens to be in the same units that the end user thinks in.  To separate the concerns of data vs. user interface a new method setLengthUnitAndDisplayUnit(lengthUnitValue, lengthUnitDisplayValue) was created which takes two parameters: lengthUnitValue which indicates which units the data is in and lengthUnitDisplayValue which indicates which units the end user would like to see the measurements displayed in.  The legacy setLengthUnit() will still operate as it always has.