Fixes to animation
My changes for making mins and maxes more flexible and organized,
missed a couple of places. Fix those.

Put elevationWindow into animationEngine, as activeWindow is. This
fixes some inconsistencies.

Put initial values for activeWindow and elevationWindow as an argument
to the AnimationEngine constructor. Take those values, and shove them
in the playbar DOM, overriding anything in the HTML. [I'm afraid to
take them out of the HTML, but maybe someday....]

Baby steps at cleaning up the spaghetti mess of potree code.