Panorama mouseup to enter, no unintentional hover
Changed the triggering event to view a panorama to Mouseup, and added code to cancel highlighting if the user is actually dragging the view around, at which point a hover is undesirable.

This made the experience with the viewer a lot smoother, for example, we have a project with many panorama spheres. In its old form, clicking down to pan the view around would result in an unintentional "focus" on the panorama image. With this new method, the user has to very explicitly want to focus on the image. 

Also seems to prevent a few other strange bugs that would happen when accidentally focusing into an image, making the pano spheres disappear and the red "hover" sphere being made a permanent background after exiting the pano view.